Merge jm/fix-draw-order into master: compute the correct draw order of game objects via draw graph.
Also make improvements to draw order calculation.

TODO: had to cut different camera orientations, not sure whats wrong. Looks like it might be a problem with finding the origin.

There is an apparent bug where the draw order looks wonky if gameobjects are on uneven terrain or if one is passing through the other. Not worried about this. Need to ensure that gameobjects always spawn on a flat surface, and need to modify astar to disallow passthrus.